+++ src/component/events.js
@@ -1,3 +1,5 @@
+const deprecate = require('../utils/deprecate');
+
 function callBeforeCreate(context) {
     if(typeof context.onBeforeCreate === 'function'){
         return context.onBeforeCreate.call(context);
@@ -12,7 +14,14 @@ function callCreate(context) {
 
 function callRender(context) {
     if(typeof context.onRender === 'function'){
+        deprecate.once('onRender is deprecated since v. 1.0.0, use onMount instead');
         context.onRender.call(context);
+    }
+}
+
+function callMount(context) {
+    if(typeof context.onMount === 'function'){
+        context.onMount.call(context);
     }
 }
 
@@ -25,6 +34,12 @@ function callBeforeUpdate(context, changes) {
 function callUpdate(context, changes) {
     if(typeof context.onUpdate === 'function'){
         context.onUpdate.call(context, changes);
+    }
+}
+
+function callUnmount(context) {
+    if(typeof context.onUnmount === 'function'){
+        context.onUnmount.call(context);
     }
 }
 
@@ -45,8 +60,10 @@ module.exports = {
     callBeforeCreate,
     callCreate,
     callRender,
+    callMount,
     callBeforeUpdate,
     callUpdate,
+    callUnmount,
     callBeforeDestroy,
     callDestroy
 };