+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.0] - 2018
+- **Added** new option property `autoCreateChildren` to component definition
+- **Added** new property `rawChildren` to component
+- **Added** new hook `onMount`
+- **Added** new hook `onUnmount`
+- **Deprecated** hooks `onRender` in favor of `onMount`
+
 ## [0.2.0] - 2018-05-23
 - **Added** "changes" param to `onBeforeUpdate` and `onUpdate`